requirements.txt
pandas
scikit-learn
shap
kafka-python
jupyter
